sow-s-strategy-bot/requirements.txt
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
aiogram>=3.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
websockets>=11.0.0
plotly>=5.0.0
